+++ core/model/diagnostic.go
@@ -35,7 +35,7 @@ type SCTDataReply struct {
 	BlockNo           uint64
 	BlockId           string
 	SmartContractData string
-	Epoch             int            
+	Epoch             uint64            
 }
 
 type NFTData struct {
